server/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
langchain
langgraph
langchain-groq
pinecone-client
httpx
mcp
langchain-community
gitpython
sentence-transformers
chardet
